Удален Main.storyboard
