+++ Tracker/Tracker/AppDelegate.swift
@@ -7,31 +7,38 @@
 
 import UIKit
 
-@main
+@UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
+    var window: UIWindow?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        
+        window = UIWindow()
+        window?.makeKeyAndVisible()
+        window?.backgroundColor = .systemBackground
+
+        let tabBarController = UITabBarController()
+        tabBarController.tabBar.barStyle = .default
+        tabBarController.tabBar.isTranslucent = false
+        tabBarController.tabBar.layer.borderColor = UIColor.ypGray.cgColor
+        tabBarController.tabBar.layer.borderWidth = 1
+
+        let trackersListViewController = TrackersViewController()
+        let trackersTabBarImage = UIImage(named: "record.circle.fill") ?? UIImage()
+        trackersListViewController.tabBarItem = UITabBarItem(title: "Трекеры",
+                                                             image: trackersTabBarImage,
+                                                             selectedImage: nil)
+
+        let statisticsViewController = StatisticsViewController()
+        let statisticsTabBarImage = UIImage(systemName: "hare.fill") ?? UIImage()
+        statisticsViewController.tabBarItem = UITabBarItem(title: "Статистика",
+                                                           image: statisticsTabBarImage,
+                                                           selectedImage: nil)
+
+        tabBarController.setViewControllers([trackersListViewController, statisticsViewController], animated: true)
+        window?.rootViewController = tabBarController
 
         return true
     }
-
-    // MARK: UISceneSession Lifecycle
-
-    func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
-        // Called when a new scene session is being created.
-        // Use this method to select a configuration to create the new scene with.
-        let sceneConfiguration = UISceneConfiguration(name: "Main",
-                                                      sessionRole: connectingSceneSession.role)
-        sceneConfiguration.delegateClass = SceneDelegate.self
-        return sceneConfiguration
-    }
-
-    func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
-        // Called when the user discards a scene session.
-        // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
-        // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
-    }
-
-
 }
